Message about incorrect operation id

+++ src/database.py
@@ -124,13 +124,19 @@ def delete_operation(telegram_address, operation_id):
     """Delete operation with current id.
 
     :param operation_id: id of operation.
+    return True if successful and False otherwise.
     """
     session = sessionmaker(bind=engine)()
     # SQLAlchemy Query object (contains db response)
     q = session.query(Operations).get(operation_id)
+
+    success = False
     if q is not None and q.telegram_address == telegram_address:
         session.delete(q)
+        success = True
+
     session.commit()
+    return success
 
 
 def get_list_of_operations(telegram_address):
